+++ libvast/vast/format/syslog.hpp
@@ -35,12 +35,11 @@
 #include <type_traits>
 #include <utility>
 
+/// This namespace includes parsers and a reader for the Syslog protocol
+/// as defined in [RFC5424](https://tools.ietf.org/html/rfc5424).
 namespace vast::format::syslog {
 
-/// This namespace includes the syslog message defintion of the syslog protocol
-/// defined in [RFC 5424](https://tools.ietf.org/html/rfc5424).
-
-// maybe nil parser
+/// A parser that parses an optional value whose nullopt is presented as a dash.
 template <class Parser>
 struct maybe_nil_parser : parser<maybe_nil_parser<Parser>> {
   using value_type = typename std::decay_t<Parser>::attribute;
@@ -63,24 +62,27 @@ struct maybe_nil_parser : parser<maybe_nil_parser<Parser>> {
   Parser parser_;
 };
 
+/// Wraps a parser and allows it to be nil.
+/// @relates maybe_nil_parser
 template <class Parser>
 auto maybe_nil(Parser&& parser) {
   return maybe_nil_parser<Parser>{std::forward<Parser>(parser)};
 }
 
-// header of syslog message
+/// A Syslog message header.
 struct header {
   uint16_t facility;
   uint16_t severity;
   uint16_t version;
-  caf::optional<time> ts; // vast time
+  caf::optional<time> ts;
   std::string hostname;
   std::string app_name;
   std::string process_id;
   std::string msg_id;
 };
 
-// The header_parser parses the header of a Syslog message
+/// Parser for Syslog message headers.
+/// @relates header
 struct header_parser : parser<header_parser> {
   using attribute = header;
 
@@ -88,9 +90,8 @@ struct header_parser : parser<header_parser> {
   bool parse(Iterator& f, const Iterator& l, Attribute& x) const {
     using parsers::print, parsers::rep;
     auto is_prival = [](uint16_t in) { return in <= 191; };
-
-    // retrieve facillity and severity from prival
     auto to_facility_and_severity = [&](uint16_t in) {
+      // Retrieve facillity and severity from prival.
       if constexpr (!std::is_same_v<Attribute, unused_type>) {
         x.facility = in / 8;
         x.severity = in % 8;
@@ -108,7 +109,6 @@ struct header_parser : parser<header_parser> {
     auto timestamp = maybe_nil(parsers::time);
     auto p = pri >> version >> ' ' >> timestamp >> ' ' >> hostname >> ' '
              >> app_name >> ' ' >> process_id >> ' ' >> msg_id;
-
     if constexpr (std::is_same_v<Attribute, unused_type>)
       return p(f, l, unused);
     else
@@ -117,26 +117,25 @@ struct header_parser : parser<header_parser> {
   }
 };
 
-// one parameter (key and value) of one structured data element
+/// A parameter of a structured data element.
 using parameter = std::tuple<std::string, std::string>;
 
-// The parameter_parser parses one structured data paramter
+/// Parser for one structured data element parameter.
+/// @relates parameter
 struct parameter_parser : parser<parameter_parser> {
   using attribute = parameter;
 
   template <class Iterator, class Attribute>
   bool parse(Iterator& f, const Iterator& l, Attribute& x) const {
     using parsers::print, parsers::rep, parsers::ch;
-
-    // space, =, ", and ] are not allowed in the key of the parameter
+    // space, =, ", and ] are not allowed in the key of the parameter.
     auto key = rep(print - '=' - ' ' - ']' - '"', 1, 32);
-    // \ is used to escape specific character
+    // \ is used to escape characters.
     auto esc = ignore(ch<'\\'>);
-    // ], ", \ need to be escaped
+    // ], ", \ must to be escaped.
     auto escaped = esc >> (ch<']'> | ch<'\\'> | ch<'"'>);
     auto value = escaped | (print - ']' - '"' - '\\');
     auto p = ' ' >> key >> '=' >> '"' >> *value >> '"';
-
     if constexpr (std::is_same_v<Attribute, unused_type>)
       return p(f, l, unused);
     else
@@ -144,10 +143,10 @@ struct parameter_parser : parser<parameter_parser> {
   }
 };
 
-// all parameters of one structured data element
+/// All parameters of a structured data element.
 using parameters = vast::map;
 
-// The parameters_parser parses the parameters of a structured data element
+/// Parser for all structured data element parameters.
 struct parameters_parser : parser<parameters_parser> {
   using attribute = parameters;
 
@@ -164,18 +163,17 @@ struct parameters_parser : parser<parameters_parser> {
   }
 };
 
-// id and paramters of one structured data element
+/// A structured data element.
 using structured_data_element = std::tuple<std::string, parameters>;
 
-// The structured_data_element_parser parses one structured data element
-struct structured_data__element_parser
-  : parser<structured_data__element_parser> {
+/// Parser for structured data elements.
+/// @relates structured_data_element
+struct structured_data_element_parser : parser<structured_data_element_parser> {
   using attribute = structured_data_element;
 
   template <class Iterator, class Attribute>
   bool parse(Iterator& f, const Iterator& l, Attribute& x) const {
     using parsers::print, parsers::rep;
-
     auto is_sd_name_char = [](char in) {
       return in != '=' && in != ' ' && in != ']' && in != '"';
     };
@@ -191,20 +189,19 @@ struct structured_data__element_parser
   }
 };
 
-// all structured data elements
+/// Structured data of a Syslog message.
 using structured_data = vast::map;
 
-// The structured_data_parser parses the structured_data elements of one Syslog
-// message
+/// Parser for structured data of a Syslog message.
+/// @relates structured_data
 struct structured_data_parser : parser<structured_data_parser> {
   using attribute = structured_data;
 
   template <class Iterator, class Attribute>
   bool parse(Iterator& f, const Iterator& l, Attribute& x) const {
     using namespace parsers;
-
     auto sd
-      = structured_data__element_parser{}->*[&](structured_data_element in) {
+      = structured_data_element_parser{}->*[&](structured_data_element in) {
           if constexpr (!std::is_same_v<Attribute, unused_type>) {
             auto& [key, value] = in;
             x[key] = value;
@@ -215,10 +212,11 @@ struct structured_data_parser : parser<structured_data_parser> {
   }
 };
 
-// the message content of the syslog message
+/// Content of a Syslog message.
 using message_content = std::string;
 
-// The message_content_parser parses the message component of a Syslog message
+/// Parser for Syslog message content.
+/// @relates message_content
 struct message_content_parser : parser<message_content_parser> {
   using attribute = message_content;
   template <class Iterator, class Attribute>
@@ -226,21 +224,20 @@ struct message_content_parser : parser<message_content_parser> {
     using namespace parsers;
     using namespace parser_literals;
     auto bom = "\xEF\xBB\xBF"_p;
-    auto p = (bom >> +any) | (+any) | eoi;
-
+    auto p = (bom >> +any) | +any | eoi;
     return p(f, l, x);
   }
 };
 
-// header, data and msg describe one syslog message
+/// A Syslog message.
 struct message {
-  header header;
+  header hdr;
   structured_data data;
   caf::optional<message_content> msg;
 };
 
-// the message parses uses the header_parser, structured_datas_parser and the
-// message_content_parser to parse one Syslog message
+/// Parser for Syslog messages.
+/// @relates message
 struct message_parser : parser<message_parser> {
   using attribute = message;
 
@@ -252,11 +249,11 @@ struct message_parser : parser<message_parser> {
     if constexpr (std::is_same_v<Attribute, unused_type>)
       return p(f, l, unused);
     else
-      return p(f, l, x.header, x.data, x.msg);
+      return p(f, l, x.hdr, x.data, x.msg);
   }
 };
 
-// A reader for Syslog messages.
+/// A reader for Syslog messages.
 class reader : public multi_layout_reader {
 public:
   using super = multi_layout_reader;
@@ -293,4 +290,5 @@ private:
   type syslog_rfc5424_type_;
   type syslog_unkown_type_;
 };
-}; // namespace vast::format::syslog
+
+} // namespace vast::format::syslog